#84 - Create fetching details for user's tournament (#39)
#84 - moved saving tournaments from controller to proper service
#84 - removed redundant front-end classes
#84 - started creating api for fetching user details
#84 - added id for result
#84 - added back-end classes for task details
#84 - created first version of service for fetching details for task
#84 - moved saving tournaments from controller to proper service
#84 - removed redundant front-end classes
#84 - started creating api for fetching user details
#84 - added id for result
#84 - added back-end classes for task details
#84 - created first version of service for fetching details for task
#84 after dummy ij merge
#84 - improved test data loader for tournament
#84 - improved fetching tournament details for user
#84 - unified number types
#84 - adapt front for new back model (task details)
#84 - unified structures to db (ints instead of long where needed)
#84 - added dummy results for one tournament
#84 - added id generator for test data
#84 - moved tournament details front files to more specific folder
#84 - removed 'chosen answers' field from result
#84 - removed unused exception
#84 - created separated csv for results for one specified tournament
#84 - added tests for UserTaskDetailsService
#84 - created tests for TournamentDetailsService
#84 - moved modules mocks to specified package
#84 - updated non-working test
#84 - Kotlin some refactoring

+++ e-arbiter-tournament/src/main/java/pl/cyganki/tournament/model/QuizTask.java
@@ -20,7 +20,7 @@ public class QuizTask extends Task {
     private List<Question> questions;
 
     @Override
-    public long getMaxPoints() {
+    public int getMaxPoints() {
         return questions.size();
     }
 }
